Updated docker-compose-backup.sh
Changed varaible name from tmp_parent_backup_dir to tmp_data_dir, added making sure project_backup_dir exists before starting archive process

+++ files/docker-compose-backup.sh
@@ -11,9 +11,9 @@ backup_time=$(date +"%y%m%dT%H%M")
 project_dir="${1:-$PWD}"
 project_name=$(basename "$project_dir")
 data_dir="/var/opt/data"
+tmp_data_dir=$(mktemp -d -p "$data_dir")
+tmp_backup_dir="$tmp_data_dir/$backup_time"
 project_backup_dir="$data_dir/backups/$project_name"
-tmp_parent_backup_dir=$(mktemp -d -p "$data_dir")
-tmp_backup_dir="$tmp_parent_backup_dir/$backup_time"
 
 # exit if not a docker compose project
 if [ -f "$project_dir/docker-compose.yml" ]; then
@@ -33,10 +33,11 @@ if [ -d "$tmp_backup_dir" ]; then
   echo "[+] Backing up $project_name project to $tmp_backup_dir"
 else
   echo "[x] Could not create temporary directory ($tmp_backup_dir) for $project_name project backup"
-  [ -d $tmp_parent_backup_dir ] && rmdir $tmp_parent_backup_dir
+  [ -d $tmp_data_dir ] && rmdir $tmp_data_dir
   exit 1
 fi
 
+# backup docker-compose.yml
 echo "    - Saving docker-compose.yml config"
 cp "$project_dir/docker-compose.yml" "$tmp_backup_dir/docker-compose.yml"
 
@@ -138,11 +139,21 @@ done
 # docker compose unpause
 
 # archive backup
+retval=1
 archive_name=${project_name}_container_backup_day${backup_day}.tgz 
-[ -f "$project_backup_dir/$archive_name" ] && rm -f "$project_backup_dir/$archive_name"
-echo "[*] Compressing backup folder to $archive_name"
-tar -zcf "$project_backup_dir/$archive_name" --totals -C "$tmp_parent_backup_dir" "$backup_time"
-retval=$?
+if [ -d "$project_backup_dir" ]; then
+  [ -f "$project_backup_dir/$archive_name" ] && rm -f "$project_backup_dir/$archive_name"
+else
+  mkdir -p "$project_backup_dir"
+fi
+if [ -d "$project_backup_dir" ]; then
+  echo "[*] Compressing backup folder to $archive_name"
+  tar -zcf "$project_backup_dir/$archive_name" --totals -C "$tmp_data_dir" "$backup_time"
+  retval=$?
+else
+  echo "[x] $project_name project backup dir to archive backup does not exist and could not be created."
+  exit 1
+fi
 
 # change owner/mode of archive and remove temporary backup directory
 if [ $retval -ne 0 ]; then
@@ -151,7 +162,7 @@ if [ $retval -ne 0 ]; then
 else
   [ -f "$project_backup_dir/$archive_name" ] && chown ${CONTAINER_USER_ID:-0}:${CONTAINER_GROUP_ID:-0} "$project_backup_dir/$archive_name" && chmod 0640 "$project_backup_dir/$archive_name"
   echo "[√] Finished backing up $project_name to $archive_name."
-  [ -d "$tmp_parent_backup_dir" ] && rm -Rf "$tmp_parent_backup_dir"
+  [ -d "$tmp_data_dir" ] && rm -Rf "$tmp_data_dir"
 fi
 
 exit 0
